VAL_RESET_HEADER => RESET_VAL_HEADER
Functions which start with VAL_XXX are generally extractive, and it
makes more sense to say RESET_VAL instead of VAL_RESET.

While a longstanding ticket was open regarding questions of what should
be capitalized or not in terms of the internal API, this is kind of
the biggest issue.

It also renames VAL_RESET...which would reinitialize the payload and
extra fields beyond just the header...to RESET_VAL_CELL.  This makes
it more clear that fields beyond the header are being overwritten.