More libRebol %host-main, simplify boot extensions, rebNot
R3-Alpha's %host-main.c was originally not able to do very much,
because it used the RL_API and RXIARGs.  It could evaluate strings and
little more.  To advance the state of the console it had to be linked
to the internal API, which exposed it to issues like PUSH_GUARD_SERIES
and other core concerns.

Now that libRebol is more capable this pushes %host-main.c to be
still more like an "easy" Rebol client.  Rather than use internal
scanning this now uses transcode, and takes advantage of rebRepossess
to avoid creating an extra copy of the decompressed host source to
do so.  It also uses rebRun() functions instead of the REBARR* based
calls, and eliminates DECLARE_VALUE() and PUSH_GUARD_XXX

Also makes a simplification to the boot extension code, to get rid
of a macro used only once, and to use API REBVAL*.