Add `local_id` to port spec for UDP ports, fix Tello demo
Some applications respond to UDP packet receipt by sending back a
communication to the outbound port used to send the packet.  The only
way to keep this number from being essentially random is to use bind()
on the UDP socket prior to the first use of sendto():

https://stackoverflow.com/q/9873061/

The port spec object currently contains a `host` and a `port-id`.  If
writing a UDP listener, it would be the port to listen on.  If writing
a UDP sender, it would be the remote port to be writing to.  There was
nowhere to specify a local port ID for UDP senders.

This adds a LOCAL-ID for that purpose of use in bind(), which was not
very tricky.  The harder issue is that if one wishes to be listening
to the signals coming back *to* that port, it must be done with the
same socket...since two sockets can't bind to the same port.  The
organization of the network code was not there for supporting a socket
that could both listen -and- write.

So this commit tricks ports which specify a local-id into running code
paths for both setting up listening and sending outbound messages.
It's for a fairly narrow UDP application, so it tries to target that
with as little change as possible--pending a greater review of the
port model.

Speaking of which: R3-Alpha did not lay out the invariants of the port
model, or what datatypes it would accept at what levels.  STRING! was
allowed to make it to the WRITE part of the TCP/UDP actor, and the
underlying bytes could be wide characters or Latin1 without the user
having knowledge of which.  Yet it would write the string bytes raw
either way, giving effectively random behavior.

With Latin1-Nowhere, this bug became obvious.  For the interim this
still allows it, and does a UTF-8 conversion.  But the port model needs
a top to bottom review of what types are accepted where and why.