Fix GC problem in DOES, MAKE DATE!, unevaluated flag
DOES creates an exemplar frame and fills it (as with SPECIALIZE).

Unlike SPECIALIZE, it failed to guard the exemplar from GC during the
evaluations used to fill it.

Also fixes an assertion triggered when the seconds were being set
in a date that hadn't been marked with the flag signifying that it
had a time component...and fixes a case where running branches did
not clear the unevaluated flag on the output cell.