COMPRESS/DECOMPRESS => GZIP/GUNZIP + INFLATE/DEFLATE
Rebol's historical COMPRESS function produced a BINARY! which gave back
the default product of zlib (which is DEFLATE-format data, plus a two
byte header, and a 4-byte ALDER32 CRC) *plus* a 32-bit size in the
4 bytes at the end.

(Note: Knowing the size in advance is not *necessary* to successfully
INFLATE data...but it helps to create an exactly-sized target buffer
in advance, which avoids reallocations and/or wasted space.)

While not a particularly complex "file format", it still produced a
bit pattern that couldn't be processed by other tools.  The gzip
format--which is necessary to support for reasons of decoding http
payloads that have been compressed--also contains a 32-bit size and
is needed in the box as well.  It makes more sense if a length-encoded
format is needed to use that for consistency.

Also, R3-Alpha's lack of exposure of plain DEFLATE without the zlib
header meant processing zip files had to strip that information off,
such as this code from %unzip.reb off rebol.org:

    compressed-data: compress data
    data: copy/part
        skip compressed-data 2
        skip tail compressed-data -8

Ren-C needs to export the basic services of its embedded zlib through
the libRebol API, to avoid things like having a PNG decoder linking in
its own copy of zlib.  (This happened when Saphirion added LodePNG,
and was ultimately corrected in Ren-C.)  The details of having one
more "file format" that only Rebol supports made that quirky too.

This goes ahead and exports INFLATE and DEFLATE as raw operations
suitable for use in zip processing, and gives an /ENVELOPE refinement
for adding a Zlib or Gzip envelope.  The inflation process is given
a DETECT refinement, which can discern gzip and zlib envelopes, but
also this choice is used to finesse a technical issue of untrustworthy
sizes in some GNU-created "gzip" files...DETECT will not use the size:

http://stackoverflow.com/a/9213826

(Note: To decompress a legacy binary from R3-Alpha or Rebol2, do a
`inflate/part/envelope data (skip tail of data -4) 'zlib`.)

This commit also removes the very generically named "COMPRESS"
operation, which due to its generality glosses over the extremely
important question of what format the compressed data is put in.
*(End users can say `compress: gzip | uncompress: gunzip` if they
want to, but it doesn't seem advisable.)*

Then specializations are created for GZIP, GUNZIP, ZDEFLATE, ZINFLATE
operations.  Also those are given as API entry points, which
are separated for rebInflate()/rebDeflate() and rebGzip()/rebGunzip()
and rebZinflate()/rebZdeflate().

Also includes light cleanup and fixes for ZIP and UNZIP.