Use zlib's CRC32 table, vs. creating another one
R3-Alpha had two CRC tables, one for a 24-bit CRC and one for CRC32.

Zlib has two checksums, Adler32 and CRC32.  The CRC32 was identical
to the one implemented separately in Rebol.  However, zlib exports
the table...so it is part of the API.  This switches to use that
table instead, deleting the Rebol one.

The rationale for the 24-bit CRC is not explained, but presumably it
is there because it is faster than CRC32.  Performance should be
examined and decided if Adler32 would work as well or better, but
that is left for a future commit.