Fix infinite loop cases in FOR
Once consequence of the change suggested in #1993 was supposed to be
that it was not possible to pick a combination of start/end/bump in FOR
that would infinite loop.  However, the implementation omitted a
necessary check of the bump value.

Includes changes to use of FOR on a series and REPEAT to honor the
new semantics.