Get rid of ASSERT_NO_RELATIVE check
In the early development of specific binding--when it was still a
speculative feature--the data typing of REBVAL and RELVAL were being
split up.  The mechanics evolved to be able to use the C++ build to
do fairly good static typing to ensure there weren't mistakes.  But
a debug check for -every Init_Any_Array() call- would deep walk the
array to make sure it wasn't typed specific yet have relative values.

In over two years this expensive assertion hasn't fired once.  It is
reasonably covered by other checks.