Eliminate /CHECKSUM module option
The somewhat complex module code included an option which would
calculate a "secure" checksum of a module's body when saving it.  The
saved file's header would then contain a `checksum: #{...}` binary.
The LOAD process would verify that this checksum matched the payload
of the body data.

Such a feature is obviously not security-oriented, as anyone who has
the ability to modify the body also has the ability to modify the
header checksum.  Yet comments indicate security was a motivation:

    ; The checksum applies to the uncompressed binary source of the
    ; body, and is calculated in LOAD-HEADER if the 'checksum header
    ; field is set.  A copy of the checksum is saved in the system
    ; modules list for security.

...which is pretty obviously wrong.  Secure checksumming of a script
can't put the checksum in the script itself.

(Note: this should not be confused with the idea of embedding a
hash of *another* script, e.g. in a NEEDS: statement.  But that hash
should be of the whole file including the header.  Otherwise it could
just be changed to have its own NEEDS: statement that would import an
untrusted dependency!)

In trying to think of the use for a checksum of a file in the file
itself (though obviously only part of it) there are two *potential*
applications:

PAYLOAD ERROR DETECTION:
When it comes to error detection, the most relevant scenario would be
making sure that the compression algorithm used decompresses the data
to the original intended script data.  However, since the compression
used for the script bodies (now GZIP) has a CRC32 in it, there is
very little point in adding another hash for that purpose.  (Detecting
bit flips from solar rays is another area that can likely be taken
off the table as an important consideration in this project.  :-P)

POOR MAN'S MODIFICATION TRACKING:
A fanciful feature might be to allow someone who was not using Git or
have access to a networking diff tool to keep track of local changes
they make to scripts.  The idea would be that the checksum would be
there to remind such users of modifications made to assist in debugging
so they didn't forget they had done so.  It wouldn't help them revert
back to before the changes were made to the exact bytes of the file
that were checksummed--but it could at least remind them (or others)
they were dealing with modified files, to help contain those changes.

With infinite time, that sounds like it might be a quirky feature
to pursue.  But it would require more code and settings (the checksum
feature currently refuses to run scripts that don't match--it would
need to be a "you've changed it, keep running? [Y/N]" option to be
of any use).  The situation of monitoring local changes is better
handled by git--which would also track header changes, and be able to
put the file back to its original state after local hacking.

Rather than maintain tests of the checksum feature, this eliminates it,
with a resulting step of simplifying the module code noticeably.

Also fixes issue that NOT of a bitset was being ignored by MAKE BITSET!
which was messing up whitespace skipping before compressed payloads
in modules, and changes the load/save decompression tests to account
for the fact that not all platforms produce the same compressed output
for the same input data.