Automatically lock series backing map keys when inserted
R3-Alpha's MAP! would give wrong answers if a key used in a map were
changed after the insertion.  Ren-C safeguarded against this by
only allowing keys that had been LOCK'd immutably, with a promise
that they would remain immutable (as there is no UNLOCK).  This also
allowed ANY-ARRAY!s or OBJECT!s to be used as map keys.

For efficiency, the key passed in was not automatically copied before
locking, as that could have surprising memory usage and performance
consequences.  Instead, an already locked key would be used directly,
and a key that hadn't been locked would give an error...which would
allow the person doing the insertion to decide if they felt they
needed a separate mutable copy (shallow or deep)...and they could
either make a copy and lock it, or just lock what they had.

This commit switches the situation so that the map just locks any
unlocked keys it is passed, and puts a notice on them that they were
locked in order to be map keys.  So when a later modification is
attempted, the user will get feedback of the reason for the locking.
(The feedback is primitive and could be made more informative in the
future.)

The advantage of this is that casual users of maps may not even notice
the locking, if they weren't intending on modifying keys in the first
place.  The trouble comes up with cases like this:

    m: make map! []
    s: copy ""
    repeat n 10 [
       append s "a" ;-- says s was locked for map, 2nd time thru loop
       m/(str): n
    ]

The user would instead have to do `m/(copy str): n`.  It seems that on
balance, it's better to bias it this direction.