Fix REWORD/CASE
REWORD was changed to use a MAP! for its substitutions, even if the
caller passed in a BLOCK!.  But differently-cased copies of the same
key can no longer be accessed with plain SELECT--they'll cause an
error unless /CASE is used.

This changes that.  Also adds a more informative message in SER_AT_RAW
when the request is for a freed series (width 0), and updates the
r2r3-future.r to restore UNTIL in older Ren-Cs (which had deprecated
it in favor of LOOP-UNTIL)

Shortens some comments in Do_Core().