Eliminate REBUPT/REBIPT, standards-breaking Reb_All payload
Given the use of a `<stdint>` shim, the REBUPT and REBIPT types add
no semantic value to a declaration over `uintptr_t` and `intptr_t`.
In fact, the cases that use them just had to explain that "REBUPT is
the Rebol equivalent of uintptr_t" at the callsites.

This changes to use different marker types to give meaning in cases
that chose to use uintptr_t for a consistent reason, e.g. REBTCK for
measuring debug tick counts, or REBLIN for measuring line numbers...
since that's how many bits there are in the REBSER currently for line
counts.  (The latter is a good example of why an abstraction like
REBLIN is better, since it may be that 64-bit platforms would *not*
use the full uintptr_t size slot in the series for a line number, and
instead put more information there...so it's good to see the sites
that depend on it.)

Also fixes up several cases where a signed integer was being set up
through a cast of a uintptr_t of a pointer, mostly in the FFI.  These
would (potentially) lose bits out of the pointer when storing in a
signed integer in 64-bit builds.  While the goal of this commit was
not to specifically research and ensure those FFI cases are correct, at
the very least they got somewhat better.

As part of the change, this eliminates the Reb_All payload structure.
This was something R3-Alpha used to hash values without knowing
exactly what they were, the definition there was:

    // All bits of value fields:
    typedef struct Reb_All {
        REBCNT bits[3];
    } REBALL;

    #define VAL_ALL_BITS(v) ((v)->data.all.bits)

This was a broken concept to start with, because a union should only
be read through the last member assigned.  There are also several
types that don't canonize their bits in unused fields (or even
canonize in general), so hashing on the basis of the raw bits would
not be correct.

Not only that, but it used 32-bit REBCNTs even in the 64-bit build.
So the 64-bit build would not be reading 3 fields but the first and
half of the second.

The type had hung around for some time and was only left being used
to hash REB_MONEY and REB_PAIR values.  Since REB_PAIR is now
officially two full-band values (vs. a specially encoded numeric type
with different ranges and rules than either INTEGER! or DECIMAL!)
those values can just be hashed together.  This commit does that and
just reproduces more or less what the money hash was doing, but using
raw unsigned char access to do it.