Use C's iso646.h header to get `not`, `and`, `or`, etc.
Because there is no boolean type in C, the REBOOL datatype was actually
able to hold numbers other than 0 and 1.  This made some desirable
optimizations slower, because they had to go through the equivalent of
`rebool ? 1 : 0` in order to use the flag predictably in math
operations.  This is a limit of pre-1999 C, which Rebol is still
trying to be compatible with.

So tricks were used to validate the REBOOL usages through the
DEBUG_STRICT_BOOL checks.  These worked with the operators NOT() and
DID() to enforce the rule.

While being compatible with pre-1999 C may not seem all that useful
(as opposed to just doing `#include <stdbool.h>` and having a `bool`
type and forgetting about it), the NOT() and DID() were also useful
for increasing the visibility of things, as ! is hard to see, and
!! is rather cryptic for this purpose.

However, the result was that many logical operations were winding
up being passed through DID(), even if they were fairly obviously
boolean.  e.g. `REBOOL x = (1 > 2)` was not legal, and you would
have to write `REBOOL x = DID(1 > 2)`, because from C's point of view
`>` returns an integer...so there's no guarantee of safety.

This commit uses a smarter strategy, based on the idea of only doing
the DEBUG_STRICT_BOOL checks in C++ builds.  It simply defines FALSE
to be the C++ boolean literal `false`, and TRUE to be the C++ boolean
literal `true`, and then defines REBOOL as a class that can only be
initialized from an actual C++ boolean--no coercions from integer
are allowed.  This cleans up the code considerably, and means that
it wasn't necessary to define NOT or DID as functions to do the
checks...they could just be plain ! and !!  It still catches any cases
where attempts are made to initialize booleans directly from an
integer, e.g. `REBOOL x = (flags & SOME_FLAG_CONSTANT)`, which might
leave x holding a value besides 1 or 0.

Without the need for a custom NOT(), it suggests going ahead and
embracing the existing iso646.h standard, which defines `not`, `and`,
`or` etc. as alternatives to `!`, `&&`, `||`, etc.

https://en.wikipedia.org/wiki/C_alternative_tokens

These are more legible and more standardized (as they are predefined
in standard C++, and via the header in C as well).

Because of the death of NOT() and DID(), files containing those had
to be changed.  However, the goal is to not modify third party code,
so this is not a search-and-replace to use `and` instead of `&&` in
every file.  For some files where it made sense to change them, this
goes ahead and does so (avoiding files that are heavily modified on
the utf8-everywhere branch, and leaving the FFI extension to Shixin's
discretion on whether he wants to use the alternative tokens or not).