Transition from OneFunction to OneAction
Rebol2, R3-Alpha, and Red all have a variety of datatypes for holding
invokable code.  This implementation detail causes some confusion,
e.g. APPEND was an ACTION!, while IF was a NATIVE!, and only functions
made with MAKE FUNCTION!, FUNC, or FUNCTION would actually be FUNCTION!
The typeset ANY-FUNCTION! existed as an umbrella for any of these, but
also included OP!...which had different behavior.

Ren-C did a simplification on this to get just one interface type,
while the differing dispatch was taken care of in a more flexible and
efficient way than a switch() on a datatype...by actually storing a
C "dispatcher" function with each invokable value.  This meant new
and interesting types of dispatch could be accomplished without
having new types appear (e.g. fast TYPECHECKERs were introduced).

The first iteration of this picked the name FUNCTION! for the unified
class, but that wound up being an unfortunate name for the "noun" of
the class...because FUNCTION is entrenched as being a generator.  For
that and several other good reasons, another word is better...and
ACTION! seems a good short choice.

This is a fairly low-impact change for most users, as FUNCTION! and
ANY-FUNCTION? etc. can just be defined as :ACTION! and :ACTION?.  To
keep it from being too disruptive at the moment, those definitions are
kept in by default--since they aren't marked for other use.

However, it meant touching pretty much every C source file in the core.
So it was held off on for a long while, until it was pretty much
deemed certain as the right change to be making.

Having made the change, it does turn out to be a marked improvement,
bringing clarity to terminology (e.g. otherwise if someone sees a
FUNC in the source, would they think it was made with the FUNC
generator?)  Clear lines can be written like `func = Make_Action(...)`
in the FUNC generator, to convey more precisely what is happening.
Also, this helps reduce confusion over whether something is a C
function or a Rebol function, as the Rebol ones are ACTIONs.

Some related changes wound up in this commit, such as not changing to
RXT_ACTION but going ahead and getting rid of RXT_s, getting rid of
FUNC-CLASS-OF as a better implementation is considered, and cleaning
up other things so long as code was being examined.  The previous
operation known as FUNC_VALUE was changed to ACT_ARCHETYPE, to better
remind those using it that it does not contain a binding, and hence
ACT_ARCHETYPE(VAL_ACTION(action)) is a lossy operation, not returning
quite the same value back.