Make BAR! act as an invisible, instead of voiding
When BAR! was first introduced, there was not a concept of "invisible"
evaluation.  So it had to produce some evaluative product.  It made
void, which also happened to be neutral in R3-Alpha's ANY and ALL,
so that this definition could work there too:

    |: does []
    all [foo | bar]

With the introduction of invisibles, the idea that a BAR! could block
evaluation visibility without interfering with the result stream at
all becomes appealing.  This allows, for instance:

    return (1 + 2 | elide "hi") ;-- returns 3

That's not what happens with a voiding |.

Changing | to act as an invisible-class element, while still acting
as a barrier, is a bit more involved than one might think due to the
interaction with variadics.  Semantics of variadics continue to get
nailed down as questions arise in their design, but if they are not
hard quoting they are supposed to stop at expression barriers.  This
does some basic implementing and testing of that.

One consequence of this is that it eliminates a previous feature--which
was already in the "questionable ideas" bin--of letting BAR! serve
as a line feed in UNSPACED and PRINT.