Cut down text commentary in R3-Legacy, links to Trello
This commit cuts down the amount of expository writing in the
r3-legacy emulation functions.  It replaces any documentation of the
function or parameters with notes and links that would show up in HELP
about deprecated parameters or new features.

(The assumption being that someone who wanted to look at the HELP for
an R3-Alpha feature would have a copy of R3-Alpha in which to do so,
and it's more useful to be brief in the source to get to the point of
how the emulation works or what's been deprecated.)

Also fixes a couple of mistakes from the OneAction commit in the
r3-legacy mode.