Fix derived specifier bug, WORD! spelling failure
The LENGTH-OF an ANY-WORD! was using the API for decoding UTF-8
characters incorrectly.

Also, a case which should have been using a derived specifier was
not.  This caused an `apply 'foo/bar [...]` in a function's body
to fail to derelativize the foo/bar words in the path, if they
were words in the function.