Rebmake Simplifications
This adds some header description text for %rebmake.r, and tries to
simplify a couple of things.

One is that the FILTER-FLAG function seemed to be modifying filtered
tags on an as-needed basis, rather than having the compiler flags
bee their literal text.  This seems like it would add confusion for
those who are writing the tags, by adding another variable into
where the strings they're seeing would come from.  It does not seem
any tags were dependent on this "automatic" treatment, so seemingly
a feature that can be dropped.

It also transforms a delimiting pattern like:

    if all [D not empty? definitions] [
       unspaced [
           "/D"
           delimit map-each flg definitions [
               opt filter-flag flg id
           ]
           " /D"
        ]
    ]

Into the easier to understand:

    if D [
        spaced map-each flg definitions [
            if flg: filter-flag flg id [unspaced ["/D" flg]]
        ]
    ]