MATCH with falsey input returns BAR! success, void failure
There has been a known question about the behavior of MATCH when
its input is falsey, e.g.

    if (match [logic!] 1 > 2) [...]
    if (match [blank!] find "abc" "d") [...]

Both of these "match" but they would not run the branch, because the
result passed through is falsey.

This modifies the mechanics so that the truthful cases return BAR!,
which should hoepfully clue in the client that just because a case
passed doesn't mean the failure is giving them a "good result for
storing as pass through in a variable".  It also returns void in
the falsey cases for an even more noticeable cue.

Note: because the only return results from `match [logic!]` are BAR!
and void, they can be forced to true/false with DID.

Also includes a fix to FINISH-INIT-CORE, which as a locals gathering
PROCEDURE wasn't actually doing what it was supposed to do when
proxying lib variables into the new user context.