Retake TRY for TO-VALUE, make NULL void's API representation
This change is designed to push the "non-value" status of voids to
an even more useful level, both inside and outside of the API.

The concept is that voids represent a form of "failure" that is light,
and doesn't immediately raise an error.  It can be flipped on and off
by means of short words TRY and OPT.

Using TRY is a new idea (it had long been earmarked for being some
other error construct vs. a TRAP synonym).  Note that the concept of a
DEFANG or DEVOID operator for this type of purpose was conceived
before, and ultimately became TO-VALUE:

https://github.com/rebol/rebol-issues/issues/2003

TRY is even shorter, and pairs well with OPT.  Hence any operation
which could conceivably return any result as non-failure (including a
BLANK!) would use void as its "failure"...and BLANK! inputs to
non-mutating operations would be willing to return a void.

An interesting side-effect of this is that libRebol code using the
API can avoid the need for any rebRelease() or leaked handles in
some cases by making sure the instances it's not interested in
returns void:

    REBVAL *item = rebRun("opt match [integer! blank!] try", x, END);
    if (not item) {
        ... code that does not need rebRelease(item) ...
    }

Using NULL means being able to build on top of C's notion of truthy
and falsey, which is extremely useful.