Debug mode for tracking cells independent of payload
The ability to save file, line number, and tick information in a cell
is extremely useful.  Up until now, it has only been possible to do
this with cells that don't use their payloads (e.g. a void, BLANK!,
LOGIC!, BAR!) or that haven't used them yet.  This adds a mode that
compiles a REBVAL to 2x its ordinary size, in order to hold the
tracking information throughout the life of the value.

This caused a test of an ability that hasn't been checked for a while,
which is to be able to compile with a sizeof REBVAL that is not 4
platform pointers.  Because REBSERs were designed to be able to be
the same size as 2*REBVAL, this is a little more complicated, but
the behavior was relatively easy to switch out...so now there is
an UNUSUAL_REBVAL_SIZE compile option.