Eliminate existing rebTrap() API, tune UNEVAL/rebUneval()
In the interest of cutting down on the total number of entry points
in libRebol, this switches to using `rebRun("trap [", ...` as the
means for trapping errors in Rebol-based code.  (This may open the
door to using rebTrap() in the API specifically for when the code
being trapped is C code, vs. the rebRescue() name used now.)

The only use of rebTrap() so far had been in the call to the host
console, but changing it to use this form introduced a twist, in that
the passed parameters to the HOST-CONSOLE function were no longer in
the "outermost" scan for the rebRun() command, but inside a BLOCK! that
was spliced during the varaidic processing.  Since void elements
cannot appear in blocks, this meant the call couldn't simply pass
things like the result value directly.

Rather than avoid the problem (e.g. by switching to using something
like an APPLY), this delved into it some and rethought the definition
of UNEVAL.  This drops its previous definition as a "more generalized
QUOTE operation that could quote voids and BAR!s".  Instead it eases
off on the idea that it is all that useful to block QUOTE from quoting
BAR! (given that hard-quoting variadic operations need to anyway).

At the same time, it embraces the idea that the evaluator will not
accept feeds of data with voids in them being passed to hard quoting
slots.  Instead of somehow quoting voids to pass them to evaluative
slots, this makes UNEVAL produce a GROUP! that may contain a QUOTE or
not, and the goal is that it produce the original value back when it
is EVAL'd.  Hence a void input produces `()`, while an integer! would
produce `(quote 10)`, a GROUP! would produce `(quote (a b c))`, etc.

This would generally be useful in COMPOSE situations, for splicing a
value into a slot that is known to be going to be evaluated later, and
wanting to produce the value back.  Correspondingly, a libRebol
"instruction" of rebUneval() is added to accomplish the same effect
without a COMPOSE, just inline in the variadic arguments.