VOID? => NULL?, add nullptr shim, `;-- null` eval output
This changes void? to null?, and includes a basic "nullptr" shim to
help with bugs in passing NULL to variadics in C.  This is based on the
reasoning here:

https://forum.rebol.info/t/null-in-the-librebol-api-and-void-null/597

It also adds a commentary `;-- null` to the output when an evaluation
makes a null result.  By doing this, it helps to reinforce to the
user what they are looking at...and that this "null" is an actionable
result (most recently as a C null pointer in the API)

Proactively changing the entire source base to use nullptr instead of
NULL is not done in this step.  Just a few files related to the
evaluator.

A more interesting study is of whether to use the `if (pointer)` and
`if (not pointer)` idioms instead of `if (pointer != NULL)` and
`if (pointer == NULL)`.  The question would be if it cleans the code
up enough to warrant not annotating at the reference site that the
thing you are talking about is a pointer instead of a boolean.  It
seems there may not be a 100% rule on this.  But since in the Rebol
language itself there's a much more fuzzy use of states of non-LOGIC!
things to do branching...that it would seem the interpreter source
itself should be willing to embrace it, if the result looks cleaner.

Those experiments aside, the important matter of committing to the
API notion of NULL suggests syncing up the name.