Add new MATCH mechanic for "stealing" a filter function's arg
Previously there had been some question about how the MATCH function
might be able to do more than just match datatypes.  Should there be
a "matching dialect", that would pass through the value if it matched?
Yet defining a complex dialect seemed to overlap with PARSE.

This commit uses a better plan: to borrow the kind of approach that
DOES uses for specializing an action from its invocation, to re-use
PARSE (or any other ACTION! one uses for invocation).  It "steals"
the first argument from the invocation frame, while using the result
as the filter signal:

    >> match parse "aaa" [some "a"]
    == "aaa" ;-- ta-da!

    >> match parse "aaa" [some "b"]
    == _

This allows MATCH to improve on even the single-arity function form
of EITHER-TEST, by not needing the GET-WORD!:

    >> either-test :even? 10 [print "branch" <branched>]
    == 10

    >> either-test :even? 3 [print "branch" <branched>]
    branch
    == <branched>

    >> match even? 10
    == 10

    >> match even? 3
    == _

With the pressure off of defining a "match dialect", EITHER-TEST could
be simplified and optimized...though EITHER-TEST-VALUE and
EITHER-TEST-NULL still are speedier and worth it due to their use
as ELSE and ALSO.