Eliminate FIRST+, add REDUCE/TRY and experimental REDUCE/OPT
The FIRST+ SERIES idea originated in R3-Alpha. The concept was that it
would be possible to not mutate a series, but get the value at the
current position and advance the position of the series variable.  Thi
 means it was a quoting operator, as it needed to update SERIES.

Ren-C tries not to make such operations that do not clue the reader
into the idea that a variable is modified.  For instance: instead of
`default 'x 3` it uses `x: default [3]` and quotes the left hand side,
so that the word acquisition suggests the *capability* to change X.

But beyond that, FIRST+ adds a lot of questions.  Is this a general
meaning for the suffix of "+" (if so, it seems limited).  Is there
a SECOND+, and what does it do?

It didn't sit well with everyone at the time of its introduction, and
a PR for FIRST+ to Red was rejected:

    https://github.com/red/red/pull/1105

If FIRST+ could be used properly is an open question, considering that
the three usages of it were obfuscating.  e.g. in R3-Alpha module
startup:

    code: reduce [first+ code code]

Which is in this commit made clearer with:

    code: reduce [first code next code]

Additionally this adds a /TRY refinement so that REDUCE instructions
will convert any null evaluations to blank.  The more speculative
/OPT refinement is added, which impinges a bit of COMPOSE's "turf"
by allowing null evaluations to disappear.  This is added to see if
anyone finds a good use for it vs. COMPOSE.